pyproject.toml
[project]
name = "photoprocessing-develop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "fastapi>=0.121.2",
    "openai>=2.8.0",
    "pathlib>=1.0.1",
    "pillow>=12.0.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
